# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopClone E-Commerce API",
    "description": "A full-featured e-commerce REST API powering the ShopClone storefront. Provides endpoints for user authentication, product catalog browsing, shopping cart management, order processing, wishlists, product reviews, and simulated payment processing.",
    "version": "1.0.0",
    "contact": {
      "name": "ShopClone API Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get the current user's cart",
        "description": "Retrieves all items in the authenticated user's shopping cart, including full product details and quantities.",
        "operationId": "getCart",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved cart contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                },
                "example": {
                  "items": [
                    {
                      "product": {
                        "id": "prod-001",
                        "title": "Wireless Bluetooth Headphones",
                        "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                        "price": 79.99,
                        "image": "https://example.com/images/headphones.jpg",
                        "category": "electronics"
                      },
                      "quantity": 2
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Add an item to the cart",
        "description": "Adds a product to the authenticated user's shopping cart. If the product already exists in the cart, the quantity is incremented. An optional quantity parameter can be specified (defaults to 1).",
        "operationId": "addToCart",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartRequest"
              },
              "example": {
                "productId": "prod-001",
                "quantity": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item successfully added to cart; returns the updated cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                },
                "example": {
                  "items": [
                    {
                      "product": {
                        "id": "prod-001",
                        "title": "Wireless Bluetooth Headphones",
                        "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                        "price": 79.99,
                        "image": "https://example.com/images/headphones.jpg",
                        "category": "electronics"
                      },
                      "quantity": 1
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing productId or invalid quantity)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Product ID is required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Cart"],
        "summary": "Update cart item quantity",
        "description": "Updates the quantity of a specific product in the authenticated user's cart. If the quantity is set to 0, the item is removed from the cart.",
        "operationId": "updateCartItem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartRequest"
              },
              "example": {
                "productId": "prod-001",
                "quantity": 3
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart item successfully updated; returns the updated cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                },
                "example": {
                  "items": [
                    {
                      "product": {
                        "id": "prod-001",
                        "title": "Wireless Bluetooth Headphones",
                        "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                        "price": 79.99,
                        "image": "https://example.com/images/headphones.jpg",
                        "category": "electronics"
                      },
                      "quantity": 3
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing productId or invalid quantity)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Product ID and quantity are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Clear the entire cart",
        "description": "Removes all items from the authenticated user's shopping cart.",
        "operationId": "clearCart",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart successfully cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                },
                "example": {
                  "items": []
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get an order by ID",
        "description": "Retrieves the details of a specific order belonging to the authenticated user.",
        "operationId": "getOrder",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the order",
            "schema": {
              "type": "string"
            },
            "example": "ord-abc123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleOrderResponse"
                },
                "example": {
                  "order": {
                    "id": "ord-abc123",
                    "items": [
                      {
                        "product": {
                          "id": "prod-001",
                          "title": "Wireless Bluetooth Headphones",
                          "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                          "price": 79.99,
                          "image": "https://example.com/images/headphones.jpg",
                          "category": "electronics"
                        },
                        "quantity": 2
                      }
                    ],
                    "total": 159.98,
                    "date": "2026-02-23T14:30:00.000Z",
                    "status": "processing"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Order not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Orders"],
        "summary": "Update order status (cancel, return, or replace)",
        "description": "Performs an action on an order belonging to the authenticated user. Supported actions: 'cancel' (only for pending/processing orders), 'return' (only for delivered orders), 'replace' (only for delivered orders).",
        "operationId": "updateOrderStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the order",
            "schema": {
              "type": "string"
            },
            "example": "ord-abc123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusRequest"
              },
              "example": {
                "action": "cancel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleOrderResponse"
                },
                "example": {
                  "order": {
                    "id": "ord-abc123",
                    "items": [
                      {
                        "product": {
                          "id": "prod-001",
                          "title": "Wireless Bluetooth Headphones",
                          "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                          "price": 79.99,
                          "image": "https://example.com/images/headphones.jpg",
                          "category": "electronics"
                        },
                        "quantity": 2
                      }
                    ],
                    "total": 159.98,
                    "date": "2026-02-23T14:30:00.000Z",
                    "status": "cancelled"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid action or action not allowed for the current order status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Order can only be cancelled when pending or processing"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Order not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the /api/auth/signup or /api/auth/signin endpoints. Include in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "description": "Represents a product in the catalog",
        "required": ["id", "title", "description", "price", "image", "category"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product",
            "example": "prod-001"
          },
          "title": {
            "type": "string",
            "description": "Name of the product",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product",
            "example": "Premium noise-cancelling wireless headphones with 30-hour battery life."
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price of the product in USD",
            "minimum": 0,
            "example": 79.99
          },
          "image": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/images/headphones.jpg"
          },
          "category": {
            "type": "string",
            "description": "Category the product belongs to",
            "example": "electronics"
          }
        }
      },
      "CartItem": {
        "type": "object",
        "description": "Represents an item in the shopping cart, combining a product with a quantity",
        "required": ["product", "quantity"],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product in the cart",
            "minimum": 1,
            "example": 2
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "description": "Represents an item within an order",
        "required": ["product", "quantity"],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product ordered",
            "minimum": 1,
            "example": 2
          }
        }
      },
      "Order": {
        "type": "object",
        "description": "Represents a customer order",
        "required": ["id", "items", "total", "date", "status"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the order",
            "example": "ord-abc123"
          },
          "items": {
            "type": "array",
            "description": "List of items in the order",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Total cost of the order in USD",
            "minimum": 0,
            "example": 159.98
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the order was placed",
            "example": "2026-02-23T14:30:00.000Z"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "description": "The current status of an order",
        "enum": ["pending", "processing", "shipped", "delivered", "cancelled", "return_requested", "replacement_requested"],
        "example": "pending"
      },
...
      "AddToCartRequest": {
        "type": "object",
        "description": "Request body for adding a product to the cart",
        "required": ["productId"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The unique identifier of the product to add",
            "example": "prod-001"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity to add (defaults to 1 if not specified)",
            "minimum": 1,
            "default": 1,
            "example": 1
          }
        }
      },
      "UpdateCartRequest": {
        "type": "object",
        "description": "Request body for updating a cart item's quantity",
        "required": ["productId", "quantity"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The unique identifier of the product to update",
            "example": "prod-001"
          },
          "quantity": {
            "type": "integer",
            "description": "The new quantity for the cart item. Set to 0 to remove the item.",
            "minimum": 0,
            "example": 3
          }
        }
      },
...
      "UpdateOrderStatusRequest": {
        "type": "object",
        "description": "Request body for updating an order's status via an action",
        "required": ["action"],
        "properties": {
          "action": {
            "type": "string",
            "description": "The action to perform on the order. 'cancel' cancels pending/processing orders, 'return' requests a return for delivered orders, 'replace' requests a replacement for delivered orders.",
            "enum": ["cancel", "return", "replace"],
            "example": "cancel"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message describing what went wrong",
            "example": "An unexpected error occurred"
          }
        }
      },
...
      "CartResponse": {
        "type": "object",
        "description": "Response containing the current state of the shopping cart",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "description": "Array of cart items (product with quantity)",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
...
      "SingleOrderResponse": {
        "type": "object",
        "description": "Response containing a single order",
        "required": ["order"],
        "properties": {
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token is invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Authentication required"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}